moved action properties out so other blocks can use
