+++ apps/journeys-admin/src/components/Editor/ControlPanel/Attributes/Action/Action.tsx
@@ -7,7 +7,7 @@ import InputLabel from '@mui/material/InputLabel'
 import { Typography } from '@mui/material'
 import KeyboardArrowDownRoundedIcon from '@mui/icons-material/KeyboardArrowDownRounded'
 import Stack from '@mui/material/Stack'
-import { GetJourney_journey_blocks_ButtonBlock as ButtonBlock } from '../../../../../../../../__generated__/GetJourney'
+import { GetJourney_journey_blocks_ButtonBlock as ButtonBlock } from '../../../../../../__generated__/GetJourney'
 import { NavigateNext } from './NavigateNext'
 import { NavigateStep } from './NavigateStep'
 import { NavigateJourney } from './NavigateJourney'
