+++ apps/journeys-admin/src/components/Editor/ControlPanel/Attributes/Action/NavigateJourney/index.ts
