+++ apps/journeys-admin/src/components/Editor/ControlPanel/Attributes/Action/NavigateLink/NavigateLink.spec.tsx
@@ -1,7 +1,7 @@
 import { MockedProvider } from '@apollo/client/testing'
 import { render, waitFor, fireEvent } from '@testing-library/react'
 import { EditorProvider, TreeBlock } from '@core/journeys/ui'
-import { GetJourney_journey_blocks_ButtonBlock as ButtonBlock } from '../../../../../../../../../__generated__/GetJourney'
+import { GetJourney_journey_blocks_ButtonBlock as ButtonBlock } from '../../../../../../../__generated__/GetJourney'
 import { NavigateLink, NAVIGATE_TO_LINK_ACTION_UPDATE } from './NavigateLink'
 
 describe('NavigateLink', () => {
