+++ apps/journeys-admin/src/components/Editor/ControlPanel/Attributes/Action/NavigateLink/index.ts
