+++ apps/journeys-admin/src/components/Editor/ControlPanel/Attributes/Action/NavigateNext/index.ts
