+++ apps/journeys-admin/src/components/Editor/ControlPanel/Attributes/Action/NavigateStep/index.ts
